smaller size

+++ include/ylt/metric/counter.hpp
@@ -185,11 +185,11 @@ struct array_hash {
 using counter_t = basic_static_counter<int64_t>;
 using counter_d = basic_static_counter<double>;
 
-template <typename T, size_t N>
+template <typename T, uint8_t N>
 using dynamic_metric_hash_map =
     std::unordered_map<std::array<std::string, N>, T, array_hash<N>>;
 
-template <typename value_type, size_t N>
+template <typename value_type, uint8_t N>
 class basic_dynamic_counter : public dynamic_metric {
  public:
   // dynamic labels value
@@ -378,7 +378,7 @@ class basic_dynamic_counter : public dynamic_metric {
 
   bool has_label_value(const std::vector<std::string> &label_value) override {
     std::array<std::string, N> arr{};
-    size_t size = (std::min)(N, label_value.size());
+    size_t size = (std::min)((size_t)N, label_value.size());
     for (size_t i = 0; i < size; i++) {
       arr[i] = label_value[i];
     }
